requirements.txt
requests
python-dateutil
beepy